Refactor puzzles to separate concerns
